Connect sigFlash to flash method
When using setBrush to change the color of regions of interest
during the active task, only the part of the region containing
the trace was colored. However, when a screen flash was called,
it somehow resets the plot such that the regions are colored
appropriately. Notably, it does not appear to clear/reset or
otherwise re-initialize the regions because they are grey by
default when initialized but it's possible for them to be reset
to green, which shouldn't be possible if they're being re-
initialized (i.e., the init_line_region method is called).

To avoid actually coloring the screen, None is
sent to the flash method to produce a transparent background