Stability changes
Change config times from ms to seconds for better UX
Start rest taks when start button is pressed. This change assumes
the user is starting in the base range because it will start a new
trial regardless of the location of the participant.

Previously, it would only start a trial when the user entered the
base region, but the since the user typically starts in the base
region, they would need to exit and reenter to begin a trial.

Pause randomization was reworked to be more intuitive for the user.